+++ src/cli/main.rs
@@ -49,6 +49,7 @@ where
                 Subcommand::Shutdown(cmd) => cmd.run(client).await,
                 Subcommand::Healthcheck(cmd) => cmd.run(client).await,
                 Subcommand::F3(cmd) => cmd.run(client).await,
+                Subcommand::WaitApi(cmd) => cmd.run(client).await,
             }
         })
 }